Refactor DialogOpen class into its own py file
